fix(primitives-p2p): split off p2p module
