+++ node/Cargo.toml
@@ -18,6 +18,7 @@ libp2p-length-prefix-codec = { workspace = true }
 polka-storage-proofs = { workspace = true, features = ["substrate"] }
 polka-storage-runtime.workspace = true
 primitives = { workspace = true, features = ["serde", "std"] }
+primitives-p2p = { workspace = true, features = ["serde"] }
 
 async-trait = { workspace = true }
 cbor4ii = { workspace = true }
