+++ primitives-p2p/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+license-file.workspace = true
+name = "primitives-p2p"
+repository.workspace = true
+version = "0.1.0"
+
+[features]
+serde = ["dep:serde"]
+
+[dependencies]
+cid = { workspace = true, default-features = false, features = ["alloc"] }
+ed25519-dalek = { workspace = true, features = ["pem"] }
+hex = { workspace = true }
+libp2p = { workspace = true }
+
+# Optional dependencies
+serde = { workspace = true, optional = true, features = ["derive"] }
+
+[lints]
+workspace = true
