+++ primitives/Cargo.toml
@@ -14,7 +14,6 @@ codec = { workspace = true, features = ["derive"] }
 ed25519-dalek = { workspace = true, optional = true, features = ["pem"] }
 filecoin-proofs = { workspace = true, optional = true }
 hex = { workspace = true, optional = true }
-libp2p = { workspace = true, optional = true }
 scale-decode = { workspace = true, features = ["derive"] }
 scale-encode = { workspace = true, features = ["derive"] }
 scale-info = { workspace = true, features = ["derive"] }
@@ -47,7 +46,6 @@ std = [
   "dep:ed25519-dalek",
   "dep:filecoin-proofs",
   "dep:hex",
-  "dep:libp2p",
   "scale-info/std",
   "serde/std",
   "sha2/std",
