+++ primitives/src/lib.rs
@@ -6,9 +6,6 @@ pub mod proofs;
 pub mod randomness;
 pub mod sector;
 
-#[cfg(feature = "std")]
-pub mod p2p;
-
 #[cfg(feature = "testing")]
 pub mod testing {
     // NOTE(@jmg-duarte,22/01/2025): Since there's only one thing, star import for now.
