+++ storage-provider/server/src/config.rs
@@ -8,10 +8,8 @@ use std::{
 use clap::Args;
 use libp2p::{identity::Keypair, Multiaddr, PeerId};
 use polka_storage_provider_common::config::sealing::SealingConfiguration;
-use primitives::{
-    p2p::{keypair_value_parser, validate_tcp_multiaddr, validate_ws_multiaddr},
-    proofs::{RegisteredPoStProof, RegisteredSealProof},
-};
+use primitives::proofs::{RegisteredPoStProof, RegisteredSealProof};
+use primitives_p2p::{keypair_value_parser, validate_tcp_multiaddr, validate_ws_multiaddr};
 use serde::{de::Error, Deserialize, Deserializer};
 use url::Url;
